+++ backend/commons/index.js
@@ -1,7 +1,7 @@
-'use strict';
+"use strict";
 
 // Require modules.
-const constants = require('./constants');
+const constants = require("./status-codes");
 
 // Export.
 exports.statusCodes = constants.statusCodes;
