--- backend/configurations/database.js
@@ -1,11 +1,0 @@
-'use strict';
-
-// Require modules.
-const mongoose = require('mongoose');
-
-// Try to connect with database.
-try {
-    mongoose.connect('mongodb://localhost:27017/detectionly', {useNewUrlParser: true, useUnifiedTopology: true});
-} catch (error) {
-    console.log(error);
-}
