+++ backend/initializer/index.js
@@ -1,0 +1,7 @@
+"use strict";
+
+// Require modules.
+const server = require("./server");
+
+// Export.
+exports.initializeServer = server.initializeServer;
