+++ backend/resolvers/mutation/registration.js
@@ -1,63 +1,63 @@
-'use strict';
+"use strict";
 
 // Require modules.
-const utils = require('../../utils');
-const {hash} = require('bcryptjs');
-const commons = require('../../commons');
+const utils = require("../../utils");
+const {hash} = require("bcryptjs");
+const commons = require("../../commons");
 
 // Function to register user.
 async function registerUser(userRegistrationData) {
     // Try to check for existence in database using utility function.
-    let userRecordExistsInDatabase = false;
+    let userRecordExistsInDatabase = true;
     try {
         userRecordExistsInDatabase = await utils.checkIfUserRecordExistsInDatabase(userRegistrationData.emailAddress);
     } catch (error) {
-        console.log(error);
+        console.error(error);
     }
 
-    // Register user if userRecordExistsInDatabase == false.
+    // Register user if userRecordExistsInDatabase is false.
     if (!userRecordExistsInDatabase) {
         // Try to check if all registration data is provided.
         let isAllUserRegistrationDataProvided = false;
         try {
             isAllUserRegistrationDataProvided = await utils.checkIfAllUserRegistrationDataIsProvided(userRegistrationData);
         } catch (error) {
-            console.log(error);
+            console.error(error);
         }
 
-        // If isAllUserRegistrationDataProvided == true, proceed with user registration.
+        // If isAllUserRegistrationDataProvided is true, proceed with user registration.
         if (isAllUserRegistrationDataProvided) {
             // Try to hash password.
             try {
                 userRegistrationData.password = await hash(userRegistrationData.password, 12);
             } catch (error) {
-                console.log(error);
+                console.error(error);
             }
 
             // Try to save user to database using utility function.
             try {
                 const statusCode = await utils.saveUserToDatabase(userRegistrationData);
 
-                // If successful, return response.
+                // If successful, return CREATED response.
                 if (statusCode === 201) {
-                    return ({statusCode: statusCode, statusMessage: "CREATED", responseMessage: "Registration Successful"});
+                    return {statusCode: statusCode, statusMessage: "CREATED", responseMessage: "Registration Successful"};
                 }
-                // If unsuccessful, return response.
+                // If unsuccessful, return BAD REQUEST response.
                 else {
-                    return ({statusCode: statusCode, statusMessage: "BAD REQUEST", responseMessage: "Registration Failed"});
+                    return {statusCode: statusCode, statusMessage: "BAD REQUEST", responseMessage: "Registration Failed"};
                 }
             } catch (error) {
-                console.log(error);
+                console.error(error);
             }
         }
-        // Else if isAllUserRegistrationDataProvided == false, return response.
+        // Else if isAllUserRegistrationDataProvided == false, return NOT ACCEPTABLE response.
         else {
-            return ({statusCode: commons.statusCodes["NOT ACCEPTABLE"], statusMessage: "NOT ACCEPTABLE", responseMessage: "Insufficient Registration Data Provided"});
+            return {statusCode: commons.statusCodes["NOT ACCEPTABLE"], statusMessage: "NOT ACCEPTABLE", responseMessage: "Insufficient Registration Data Provided"};
         }
     }
-    // Return conflict response if userRecordExistsInDatabase == true.
+    // Return CONFLICT response if userRecordExistsInDatabase == true.
     else {
-        return ({statusCode: commons.statusCodes.CONFLICT, statusMessage: "CONFLICT", responseMessage: "User already exists"});
+        return {statusCode: commons.statusCodes.CONFLICT, statusMessage: "CONFLICT", responseMessage: "User already exists"};
     }
 }
