+++ backend/resolvers/mutation/update.js
@@ -1,10 +1,10 @@
-'use strict';
+"use strict";
 
 // Require modules.
-const utils = require('../../utils');
-const models = require('../../models');
-const commons = require('../../commons');
-const {hash} = require('bcryptjs');
+const utils = require("../../utils");
+const models = require("../../models");
+const commons = require("../../commons");
+const {hash} = require("bcryptjs");
 
 // Function to update user personal information.
 async function updateUserInformation(userEmailAddress, updatedInformation) {
@@ -35,17 +35,17 @@ async function updateUserInformation(userEmailAddress, updatedInformation) {
     try {
         userRecordExistsInDatabase = await utils.checkIfUserRecordExistsInDatabase(userEmailAddress);
     } catch (error) {
-        console.log(error);
+        console.error(error);
     }
 
-    // Update if exists.
+    // Update if record exists.
     if (userRecordExistsInDatabase) {
         // Find, update user and return OK response.
         try {
             await models.User.findOneAndUpdate(conditions, update, options).exec();
             return ({statusCode: commons.statusCodes.OK, statusMessage: "OK", responseMessage: "Update Successful"});
         } catch (error) {
-            console.log(error);
+            console.error(error);
         }
     }
     // Else return NOT FOUND response.
@@ -61,23 +61,41 @@ async function updateUserPassword(userEmailAddress, currentPasswordEntered, newP
     try {
         userRecordExistsInDatabase = await utils.checkIfUserRecordExistsInDatabase(userEmailAddress);
     } catch (error) {
-        console.log(error);
+        console.error(error);
     }
 
     // Proceed if user exists.
     if (userRecordExistsInDatabase) {
-        // Compare entered password with correct password.
-        const comparisonResult = await utils.authenticateUser(userEmailAddress, currentPasswordEntered);
+        // Try to compare entered password with correct password.
+        let comparisonResult = false;
+        try {
+            comparisonResult = await utils.authenticateUser(userEmailAddress, currentPasswordEntered);
+        } catch (error) {
+            console.error(error);
+        }
 
-        // If comparisonResult === true, update user password, and return OK response.
+        // If comparisonResult is true, hash user password, update and return OK response.
         if (comparisonResult) {
-            try {
-                // Hash new password before update.
-                newPassword = await hash(newPassword, 12);
-                await models.User.findOneAndUpdate({emailAddress: userEmailAddress}, {password: newPassword}, {useFindAndModify:false, new: true}).exec();
-                return ({statusCode: commons.statusCodes.OK, statusMessage: "OK", responseMessage: "Password Update Successful"});
-            } catch (error) {
-                console.log(error);
+            // Check if new password is provided.
+            if (newPassword != null || newPassword !== "") {
+                // Try to hash new password.
+                try {
+                    newPassword = await hash(newPassword, 12);
+                } catch (error) {
+                    console.error(error);
+                }
+
+                // Try to update password.
+                try {
+                    await models.User.findOneAndUpdate({emailAddress: userEmailAddress}, {password: newPassword}, {useFindAndModify:false, new: true}).exec();
+                    return ({statusCode: commons.statusCodes.OK, statusMessage: "OK", responseMessage: "Password Update Successful"});
+                } catch (error) {
+                    console.error(error);
+                }
+            }
+            // Else return NOT ACCEPTABLE response.
+            else {
+                return ({statusCode: commons.statusCodes["NOT ACCEPTABLE"], statusMessage: "NOT ACCEPTABLE", responseMessage: "New Password Not Provided"});
             }
         }
         // Else return UNAUTHORIZED response.
