+++ backend/schemas/inputs.js
@@ -1,4 +1,4 @@
-'use strict';
+"use strict";
 
 // Define inputs.
 // For user registration.
